(PC-12068)[API]sendinblue transactional email - correction renaming file and tests
