+++ api/src/pcapi/admin/custom_views/support_view.py
@@ -19,9 +19,7 @@ from pcapi.admin import base_configuration
 from pcapi.connectors.beneficiaries import jouve_backend
 import pcapi.core.fraud.api as fraud_api
 import pcapi.core.fraud.models as fraud_models
-from pcapi.core.mails.transactional.users.subscription_document_error_email import (
-    send_subscription_document_error_email,
-)
+from pcapi.core.mails.transactional.users.subscription_document_error import send_subscription_document_error_email
 from pcapi.core.subscription import messages as subscription_messages
 import pcapi.core.subscription.api as subscription_api
 import pcapi.core.subscription.exceptions as subscription_exceptions
