+++ api/src/pcapi/core/mails/transactional/bookings/booking_confirmation_to_beneficiary.py
