+++ api/src/pcapi/core/mails/transactional/users/__init__.py
@@ -1,1 +1,1 @@
-from .reset_password_email import *
+from .reset_password import *
