+++ api/src/pcapi/core/mails/transactional/users/accepted_as_beneficiary.py
