+++ api/src/pcapi/core/mails/transactional/users/birthday_to_newly_eligible_user.py
@@ -9,7 +9,9 @@ from pcapi.models.feature import FeatureToggle
 from pcapi.utils.urls import generate_firebase_dynamic_link
 
 
-def get_anniversary_age_18_user_email_data(user: users_models.User) -> Union[dict, SendinblueTransactionalEmailData]:
+def get_birthday_age_18_to_newly_eligible_user_email_data(
+    user: users_models.User,
+) -> Union[dict, SendinblueTransactionalEmailData]:
     if not FeatureToggle.ENABLE_SENDINBLUE_TRANSACTIONAL_EMAILS.is_active():
         email_link = generate_firebase_dynamic_link(path="id-check", params={"email": user.email})
         granted_deposit = get_granted_deposit(user, user.eligibility)
@@ -23,9 +25,9 @@ def get_anniversary_age_18_user_email_data(user: users_models.User) -> Union[dic
             },
         }
 
-    return SendinblueTransactionalEmailData(template=TransactionalEmail.ANNIVERSARY_AGE_18.value)
+    return SendinblueTransactionalEmailData(template=TransactionalEmail.BIRTHDAY_AGE_18_TO_NEWLY_ELIGIBLE_USER.value)
 
 
-def send_newly_eligible_age_18_user_email(user: users_models.User) -> bool:
-    data = get_anniversary_age_18_user_email_data(user)
+def send_birthday_age_18_email_to_newly_eligible_user(user: users_models.User) -> bool:
+    data = get_birthday_age_18_to_newly_eligible_user_email_data(user)
     return mails.send(recipients=[user.email], data=data)
