+++ api/src/pcapi/core/mails/transactional/users/duplicate_pre_subscription_rejected.py
